Remove YAML from names, move to a package named "agent", introduce pom.xml.example and a context test
