--- examples/spring-boot-kotlin-yaml/build.gradle.kts
@@ -1,59 +1,0 @@
-plugins {
-	kotlin("jvm") version "2.2.10"
-	kotlin("plugin.spring") version "2.2.10"
-	id("org.springframework.boot") version "3.5.6"
-	id("io.spring.dependency-management") version "1.1.7"
-}
-
-group = "com.example"
-version = "0.0.1-SNAPSHOT"
-description = "An Spring Boot application with Koog AI agent configured via Spring Boot YAML"
-
-
-java {
-	toolchain {
-		languageVersion = JavaLanguageVersion.of(17)
-	}
-}
-
-repositories {
-	mavenCentral()
-}
-
-// Force all dependencies to use the same version of kotlinx-coroutines and kotlinx-serialization
-configurations.all {
-    resolutionStrategy {
-        eachDependency {
-            if (requested.group == "org.jetbrains.kotlinx" && requested.name.startsWith("kotlinx-coroutines")) {
-                useVersion("1.10.2")
-            }//https://github.com/JetBrains/koog/issues/273
-            if (requested.group == "org.jetbrains.kotlinx" && requested.name.startsWith("kotlinx-serialization")) {
-                useVersion("1.8.1")
-            }//because of two kotlinx-serialization versions in classpath (1.6.3 and 1.8.1)
-        }
-    }
-}
-
-dependencies {
-	implementation("org.springframework.boot:spring-boot-starter-web")
-	implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-	implementation("org.jetbrains.kotlin:kotlin-reflect")
-
-    implementation("ai.koog:koog-agents-jvm:0.4.2")
-    implementation("ai.koog:koog-spring-boot-starter:0.4.2")
-    implementation("aws.sdk.kotlin:s3:1.5.16")
-
-	testImplementation("org.springframework.boot:spring-boot-starter-test")
-	testImplementation("org.jetbrains.kotlin:kotlin-test-junit5")
-	testRuntimeOnly("org.junit.platform:junit-platform-launcher")
-}
-
-kotlin {
-	compilerOptions {
-		freeCompilerArgs.addAll("-Xjsr305=strict")
-	}
-}
-
-tasks.withType<Test> {
-	useJUnitPlatform()
-}
