--- examples/spring-boot-kotlin-yaml/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "spring-boot-kotlin-yaml"
