--- examples/spring-boot-kotlin-yaml/src/main/kotlin/com/example/spring_boot_kotlin_yaml/SpringBootKotlinYamlApplication.kt
@@ -1,11 +1,0 @@
-package com.example.spring_boot_kotlin_yaml
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class SpringBootKotlinYamlApplication
-
-fun main(args: Array<String>) {
-	runApplication<SpringBootKotlinYamlApplication>(*args)
-}
