--- examples/spring-boot-kotlin-yaml/src/test/kotlin/com/example/spring_boot_kotlin_yaml/SpringBootKotlinYamlApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.example.spring_boot_kotlin_yaml
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class SpringBootKotlinYamlApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
