--- examples/spring-boot-kotlin-yaml/src/test/resources/application.yml
@@ -1,22 +1,0 @@
-spring:
-  application:
-    name: spring-boot-kotlin-yaml
-
-ai:
-  koog:
-    google:
-      api-key: placeholder
-      base-url: https://generativelanguage.googleapis.com
-
-agent:
-  version: "1.0.0"
-  name: "github-assistant"
-  description: "An AI agent that can use GitHub tools."
-
-  model:
-    id: "gemini-2.0-flash-lite"
-    options:
-      temperature: 1.0
-
-  system_prompt: |
-    You are a helpful GitHub assistant.
