+++ examples/spring-boot-kotlin/src/main/kotlin/com/example/agent/SpringBootKotlinApplication.kt
@@ -1,0 +1,11 @@
+package com.example.agent
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class SpringBootKotlinApplication
+
+fun main(args: Array<String>) {
+	runApplication<SpringBootKotlinApplication>(*args)
+}
